requirements.txt
streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
lxml>=4.9.0
html5lib>=1.1
python-dotenv>=1.0.0